% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{convolve2_in_R}
\alias{convolve2_in_R}
\title{向量卷积，相当于两向量做矩阵乘法，然后沿着垂直于主对角线的方向相加。}
\usage{
convolve2_in_R(arr1, arr2, verbose = F)
}
\arguments{
\item{arr1}{int array}

\item{arr2}{int array}
}
\value{
int array
}
\description{
matrix(t(1:3), ncol=1) %*% matrix(1:4, nrow=1)
}
\examples{
convolve2_in_R(1:3, 1:4)
convolve2_in_R(1:3, 1:4, T)
}
