% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{convolve2_R}
\alias{convolve2_R}
\title{convolve of 2 vectors}
\usage{
convolve2_R(arr1, arr2)
}
\arguments{
\item{arr1}{int array}

\item{arr2}{int array}
}
\value{
int array
}
\description{
test3
}
\examples{
convolve2_R(1:3, 1:4)
convolve2_R(1:3, 1:10)
}
