% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{double_me_R}
\alias{double_me_R}
\title{Double the input number}
\usage{
double_me_R(x)
}
\arguments{
\item{x}{a number}
}
\value{
a number
}
\description{
(test3)
}
\examples{
#double_me_R(1) #看来R中整数默认是 double，要强制转换类型 as.integer()
double_me_R(as.integer(5))
}
