% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloC1}
\alias{helloC1}
\title{test3: get the length of str or str vector.}
\usage{
helloC1(greeting)
}
\arguments{
\item{greeting}{a string, or string vector}
}
\value{
the length of string or string vector
}
\description{
data type cast, AS_CHARACTER;
}
\examples{
helloC1( c("this is me!", "morning") )
}
